♻️ 시리얼라이저 상속 처리 (#8)
